Fix TPC-DS correctness, remove dead BF-operator layer, satisfy clang-tidy
Correctness (TPC-DS SF1/SF10 now run all 99 queries correctly):
- Skip RPT for scopes with a disjunctive/arbitrary join predicate
  (JoinCondition with comparison == INVALID, or ANY_JOIN). Our MemoryScan
  rewrite replaces a materialized subtree with a flat scan and cannot
  preserve the bindings such a predicate reads across the join sides
  (q13/q15/q48/q85). These fall back to plain DuckDB.
- Emit EmptyResult only from ground truth (materialized Count()==0), never
  from a flooding cardinality estimate. The estimation Bloom filter is built
  from the source table's sample, so it has false negatives; a false-zero
  estimate was wrongly rewriting a non-empty join to EmptyResult (q88).
- Floor BF-filtered sample estimates away from 0 for the same reason.
- Guard BuildMemoryScan against a bindings/types size mismatch (passthrough).

Cleanup:
- Remove the unused LogicalCreateBF/UseBF + PhysicalCreateBF/UseBF operator
  layer and its config flags (use_bitmap_filter, use_table_filter_bf,
  use_dynamic_in_filter, use_row_id_filter, skip_unfiltered_table_flag).
  The live path builds filters through TransferExecutor::BuildTransferFilters.
- Apply clang-tidy fixes required by the extension code-quality CI: remove
  const_cast, explicit narrowing casts, uint64 literals, lround.

Release build, test/sql/bloom.test (14 assertions), and full IMDB / TPC-H /
TPC-DS SF1 correctness all pass.

+++ src/predicate_transfer/table_scanner.cpp
@@ -826,8 +826,8 @@ TableScanner::CompactResult TableScanner::Compact(const vector<StatsRequest> &st
 			    max_val.type() == min_val.type()) {
 				auto &stats = result.column_stats[collected_request_indices[stats_idx]];
 				stats.has_min_max = true;
-				stats.observed_min = IntegralValue::Get(min_val).lower;
-				stats.observed_max = IntegralValue::Get(max_val).lower;
+				stats.observed_min = static_cast<int64_t>(IntegralValue::Get(min_val).lower);
+				stats.observed_max = static_cast<int64_t>(IntegralValue::Get(max_val).lower);
 			}
 		}
 	}
